refactor(NgTableParams): move scope data binding into ngTableController
Take advantage of the event channel to move `$scope` data binding over into the directive
controller where it belongs.

**WARNING**: the event `ngTableAfterReloadData` is depreciated - use
`ngTableEventChannel.onAfterReloadData` instead to subscribe to the equivalent event.

With this change, there is really no need for `NgTableParams` to keep a reference to `$scope`.

Once event removed, we can remove `$scope` from `NgTableParams`.